view profile and edit profile finished
